fix(types): map INT→Int32/BIGINT→Int64 and add widening comparisons
SQL INT is 32-bit; the lexer was mapping it to Type::Int64, which
caused incorrect column sizing and confused auto-increment narrowing.
BIGINT is now a distinct token that maps to Type::Int64.

Because INT and BIGINT now differ, WHERE expressions like
`int_col = bigint_literal` would silently return false without
widening. numeric_eq/numeric_cmp in resolve.rs handle Int32↔Int64
cross-width comparisons without touching Value::PartialEq, which
must remain strict for hash-index correctness.

fill_auto_increment_values now returns Result so out-of-range
counter values (e.g. a u64 counter that doesn't fit an Int32 column)
surface as TypeError instead of silently overflowing or truncating.

TupleSchema gains actual_serialized_size for accurate per-tuple slot
sizing once TEXT columns exist; serialized_size retains its worst-case
role for heap-page slot pre-allocation.

+++ db/tests/integration/alter_table.rs
@@ -426,7 +426,7 @@ fn add_column_with_default_new_rows_use_supplied_value() {
 
     assert_eq!(
         *rows[0].get(2).unwrap(),
-        Value::int64(7),
+        Value::int32(7),
         "supplied value must win over default"
     );
 }
